Moved `Reset Database` export into `CorePlugin`
(from `LocalBackupPlugin`)
